+++ test/autodiff.jl
@@ -5,11 +5,10 @@ include("test_utils.jl")
 rng = Random.default_rng()
 Random.seed!(rng, 0)
 
-# Test Dense Layers
 @testset "Gradient Correctness: Dense Chain" begin
     c = Chain(Dense(3, 4), Dense(4, 1))
 
-    x = randn(Float32, 3, 2)
+    x = randn(rng, Float32, 3, 2)
     ps, st = Lux.setup(rng, c)
 
     ps = ps |> Lux.ComponentArray
@@ -22,4 +21,12 @@ Random.seed!(rng, 0)
     @test gs_r == gs_z
     @test gs_r ≈ gs_fdm
     @test gs_z ≈ gs_fdm
+end
+
+@testset "Broadcasting identity custom rrule" begin
+    x = randn(rng, Float32, 3, 2)
+    gs_x_1 = Zygote.gradient(x -> sum(identity.(x)), x)[1]
+    gs_x_2 = Zygote.gradient(sum, x)[1]
+
+    @test gs_x_1 == gs_x_2
 end