+++ test/layers/recurrent.jl
@@ -5,8 +5,7 @@ include("../test_utils.jl")
 rng = Random.default_rng()
 Random.seed!(rng, 0)
 
-@testset "RNNCell" begin for rnncell in (RNNCell(3 => 5, identity),
-                                         RNNCell(3 => 5, tanh),
+@testset "RNNCell" begin for rnncell in (RNNCell(3 => 5, identity), RNNCell(3 => 5, tanh),
                                          RNNCell(3 => 5, tanh; bias=false),
                                          RNNCell(3 => 5, identity; bias=false))
     println(rnncell)