+++ test/utils.jl
@@ -79,6 +79,7 @@ end
 
     # Empty ComponentArray test
     @test_nowarn display(ComponentArray(NamedTuple()))
+    println()
 
     # Optimisers
     opt = Optimisers.ADAM(0.001f0)
@@ -86,4 +87,12 @@ end
 
     @test_nowarn Optimisers.update(st_opt, ps_c, ps_c)
     @test_nowarn Optimisers.update!(st_opt, ps_c, ps_c)
+
+    if CUDA.functional()
+        ps_c = ps_c |> gpu
+        st_opt = Optimisers.setup(opt, ps_c)
+
+        @test_nowarn Optimisers.update(st_opt, ps_c, ps_c)
+        @test_nowarn Optimisers.update!(st_opt, ps_c, ps_c)
+    end
 end